requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
groq==0.5.0
python-dotenv==1.0.1
